Feature/testing

+++ app/Http/Controllers/UserController.php
@@ -9,43 +9,59 @@ use App\Models\Game;
 use App\Http\Services\UserService;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
     //ADMIN ROUTES
     public function index(){
 
-        $allPlayers = User::all(); //object
-        $userServiceMethods = new UserService;
-        $userServiceMethods->updateWinRateAllUsers();
+        if(Auth::user()->hasRole('admin')){
+            $allPlayers = User::all(); //object
+            $userServiceMethods = new UserService;
+            $userServiceMethods->updateWinRateAllUsers();
+    
+            return response()->json(['allPlayers' => $allPlayers], 200);
+        }else if(Auth::user()->hasRole('client')){
+            return response()->json(['error' => "Unauthorized, you're not an admin"], 403);
+        }else{
+            return response()->json(['error' => "Unauthorized, you're not logged"], 401);
+        }
 
-        return $allPlayers;
     }
 
 
     //USER ROUTES
     public function show($id){
-        $games = Game::where('user_id', $id)->get();
-        return response()->json(["Your win rate" => User::find($id)['winRate'] , "Games" => $games]);
+        if(Auth::user()->id == $id){
+            $games = Game::where('user_id', $id)->get();
+            return response()->json(["Your win rate" => User::find($id)['winRate'] , "Games" => $games], 200);
+        }else{
+            return response()->json(['error' => "You can't view the games of other player"], 401);
+        }
     }
 
     public function store($id){
 
-        $user = User::find($id);
-        $userServiceMethods = new UserService;
+        if(Auth::user()->id == $id){
+            $user = User::find($id);
+            $userServiceMethods = new UserService;
 
-        $game = new Game;
-        $game['dice1'] = $game->dice1;
-        $game['dice2'] = $game->dice2;
-        $game['numberResult'] = $game->numberResult;
-        $game['textResult'] = $game->textResult;
-        $game['user_id'] = $user->id;
-        
-        $game->save();
+            $game = new Game;
+            $game['dice1'] = $game->dice1;
+            $game['dice2'] = $game->dice2;
+            $game['numberResult'] = $game->numberResult;
+            $game['textResult'] = $game->textResult;
+            $game['user_id'] = $user->id;
+            
+            $game->save();
 
-        $userServiceMethods->updateWinRate($user);
+            $userServiceMethods->updateWinRate($user);
 
-        return response()->json($game);
+            return response()->json(['game' => $game], 200);
+        }else{
+            return response()->json(['error' => "You can't view the games of other player"], 401);
+        }
     }
 
     public function update($id, Request $request){
@@ -59,17 +75,23 @@ class UserController extends Controller
         ]);
 
         if($validator->fails()){
-            return response(['error' => $validator->errors()]);
+            return response(['error' => $validator->errors()], 401);
         }else{
-            return response(['message' => "Player updated!", "Sent" => $player->update($request->all())]);
+            return response(['message' => "Player updated!", "Sent" => $player->update($request->all())], 200);
         }
     }
 
     public function destroy($id){
-        Game::where('user_id', $id)->delete();
-        User::where('id', $id)->update(['winRate' => null]);
 
-        return response(['message' => "Player Games deleted!"]);
+        if(Auth::user()->id == $id){
+            Game::where('user_id', $id)->delete();
+            User::where('id', $id)->update(['winRate' => null]);
+
+            return response(['message' => "Player Games deleted!"], 200);
+
+        }else{
+            return response(['error' => "You can't delete games of other player!"], 401);
+        }
     }
